added support for student statute

+++ Ucurricular.cpp
@@ -35,7 +35,7 @@ string Ucurricular::info()
 {
 	stringstream ss;
 	ss << codigo << "\t" << setw(9) << sigla << "\t" << nome;
-	for (int i = 0; i < 72 - nome.length(); i += 1)
+	for (size_t i = 0; i < 72 - nome.length(); i += 1)
 		ss << ' ';
 	ss << creditos << endl;
 	return ss.str();
